ENEMY ATTACKS THE PLAYER
the enemy now attacks the player holy moly that feels amazing. A simple plug n play with the scirpt with a small addition of the reset of the timer to allow multiple attacks. Flexibility is pretty darn good. The animaiton even works :0 (just won't turn off, thats the next fix)