Dungeon Diversity
Added a chest to the game, which can be opened (loot pending). Recycled the code used for the paper interaction.

The dungeon is now a lot less bare! With the newly added chest, there are new rooms where the player can find loot and fight more enemies. Spice it up a bit