Floor 2 and bug fixes!
Added a second floor!

Also fixed a bug where the player would open a chest, and also get the same interaction as if they had read the paper. Fix was to add the player's LayerMask to the Physics2D.OverLapBoxAll() function to make sure it would only work when the player was near it instead of another object