Floors Working
Floors work now. Woopee. Player moves around the scene, new dungeons are generated which work and lead to other dungeons. Clean up is nice but not 100% so will work on that.
Anvil (crafting upgrades) needs work as does not transfer from floor to floor. UI is not found which doesn't allow the player to interact. Costs also need displaying on the UI